refactor(frontend): refactor suggestions autocomplete
